用户管理对接前端

+++ pawo-power/pawo-power-web/src/main/java/com/gm/controller/TestDemo.java
@@ -8,7 +8,7 @@ import java.io.Serializable;
 @Data
 class TestDemo implements Serializable {
 
-        private int id;
+        private double id;
         private String name;
         private String sex;
         private String id_card;
